[MIG] sale_variant_configurator: Migration to 9.0
This comes from module sale_product_variants from odoomrp/odoomrp-wip.
