--- sale_product_variants/README.rst
@@ -1,41 +1,0 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-    :alt: License: AGPL-3
-
-=======================
-Sale - Product variants
-=======================
-
-This module allows you to create the product variant when a sale order is
-confirmed. It adds to the sale line a product configurator, so that selecting
-a product and its attributes can be created a new product variant.
-
-
-Usage
-=====
-
-.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
-   :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/188/8.0
-
-For further information, please visit:
-
-* https://www.odoo.com/forum/help-1
-
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/odoomrp/odoomrp-wip/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed feedback
-`here <https://github.com/odoomrp/odoomrp-wip/issues/new?body=module:%20sale_product_variants%0Aversion:%208.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-
-Credits
-=======
-
-Contributors
-------------
-* Oihane Crucelaegui <oihanecrucelaegi@avanzosc.es>
-* Pedro M. Baeza <pedro.baeza@serviciobaeza.com>
-* Ana Juaristi <ajuaristio@gmail.com>
