--- sale_product_variants/i18n/ro.po
@@ -1,111 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# * sale_product_variants
-# 
-# Translators:
-# Dorin Hongu <dhongu@gmail.com>, 2015
-msgid ""
-msgstr ""
-"Project-Id-Version: odoomrp-wip (8.0)\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-11-20 18:11+0000\n"
-"PO-Revision-Date: 2015-11-18 00:37+0000\n"
-"Last-Translator: Dorin Hongu <dhongu@gmail.com>\n"
-"Language-Team: Romanian (http://www.transifex.com/oca/odoomrp-wip-8-0/language/ro/)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Language: ro\n"
-"Plural-Forms: nplurals=3; plural=(n==1?0:(((n%100>19)||((n%100==0)&&(n!=0)))?2:1));\n"
-
-#. module: sale_product_variants
-#: view:sale.order:sale_product_variants.sale_order_extended_form_view
-msgid "(update)"
-msgstr "(actualizare)"
-
-#. module: sale_product_variants
-#: field:sale.order.line.attribute,attribute:0
-msgid "Attribute"
-msgstr "Atribute"
-
-#. module: sale_product_variants
-#: field:sale.order.line.attribute,create_uid:0
-msgid "Created by"
-msgstr "Creat de"
-
-#. module: sale_product_variants
-#: field:sale.order.line.attribute,create_date:0
-msgid "Created on"
-msgstr "Creat în"
-
-#. module: sale_product_variants
-#: view:sale.order:sale_product_variants.sale_order_extended_form_view
-msgid "Duplicate"
-msgstr "Duplică"
-
-#. module: sale_product_variants
-#: view:sale.order:sale_product_variants.sale_order_extended_form_view
-msgid "Duplicate line"
-msgstr "Duplică linia"
-
-#. module: sale_product_variants
-#: field:sale.order.line.attribute,id:0
-msgid "ID"
-msgstr "ID"
-
-#. module: sale_product_variants
-#: field:sale.order.line.attribute,write_uid:0
-msgid "Last Updated by"
-msgstr "Ultima actualizare de"
-
-#. module: sale_product_variants
-#: field:sale.order.line.attribute,write_date:0
-msgid "Last Updated on"
-msgstr "Ultima actualizare în"
-
-#. module: sale_product_variants
-#: field:sale.order.line.attribute,sale_line:0
-msgid "Order line"
-msgstr "Line comandă"
-
-#. module: sale_product_variants
-#: view:sale.order.line:sale_product_variants.sale_order_line_extended_search_view
-msgid "Product"
-msgstr "Produs"
-
-#. module: sale_product_variants
-#: field:sale.order.line,product_template:0
-msgid "Product Template"
-msgstr "Produs șablon"
-
-#. module: sale_product_variants
-#: field:sale.order.line,product_attributes:0
-msgid "Product attributes"
-msgstr "Atribute produs"
-
-#. module: sale_product_variants
-#: model:ir.model,name:sale_product_variants.model_sale_order_line
-msgid "Sales Order Line"
-msgstr "Linie comandă vânzare"
-
-#. module: sale_product_variants
-#: view:sale.order.line:sale_product_variants.sale_order_line_extended_search_view
-msgid "Template"
-msgstr "Șablon"
-
-#. module: sale_product_variants
-#: model:res.groups,name:sale_product_variants.group_product_variant_extended_description
-#: field:sale.config.settings,group_product_variant_extended_description:0
-msgid "Use extended description when having product attributes"
-msgstr ""
-
-#. module: sale_product_variants
-#: field:sale.order.line.attribute,value:0
-msgid "Value"
-msgstr "Valoare"
-
-#. module: sale_product_variants
-#: code:addons/sale_product_variants/models/sale_order.py:195
-#, python-format
-msgid "You can not confirm before configuring all attribute values."
-msgstr "Nu puteți confirma fără să specificați toate atributele"
