--- sale_product_variants/views/sale_view.xml
@@ -1,95 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <record model="ir.ui.view" id="sale_order_extended_form_view">
-            <field name="name">sale.order.extended.form</field>
-            <field name="model">sale.order</field>
-            <field name="inherit_id" ref="sale.view_order_form" />
-            <field name="priority" eval="20" />
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='order_line']/tree" position="attributes">
-                    <attribute name="editable"/>
-                </xpath>
-                <xpath expr="//field[@name='order_line']/tree//field[@name='product_id']"
-                       position="before">
-                    <field name="product_tmpl_id" />
-                </xpath>
-                <xpath expr="//field[@name='order_line']" position="attributes">
-                    <attribute name="options">{'reload_on_button': true}</attribute>
-                </xpath>
-                <xpath expr="//field[@name='order_line']/tree//field[@name='price_subtotal']"
-                       position="after">
-                    <field name="order_state" invisible="1"/>
-                    <button name="action_duplicate"
-                            state="draft"
-                            string="Duplicate"
-                            type="object"
-                            icon="STOCK_COPY"
-                            attrs="{'invisible': [('order_state', 'not in', ['draft', 'sent'])]}"
-                            help="Duplicate line" />
-                </xpath>
-                <xpath expr="//field[@name='order_line']/form//field[@name='product_id']"
-                       position="before">
-                    <field name="product_tmpl_id" />
-                    <field name="product_attribute_ids"
-                           context="{'default_owner_model': 'sale.order.line', 'default_owner_id': id}"
-                           attrs="{'invisible':[('product_attribute_ids','=',[])]}"
-                    />
-                </xpath>
-                <xpath expr="//field[@name='order_line']//form//field[@name='price_unit']"
-                       position="after">
-                    <button name="update_price_unit"
-                            states="draft" string="(update)" type="object"
-                            class="oe_edit_only oe_link" colspan="2"
-                            attrs="{'invisible':[('product_id','!=',False)]}"/>
-                </xpath>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="sale_order_line_extended_form_view">
-            <field name="name">sale.order.line.extended.form</field>
-            <field name="model">sale.order.line</field>
-            <field name="inherit_id" ref="sale.view_order_line_form2" />
-            <field name="arch" type="xml">
-                <field name="product_id" position="before">
-                    <field name="product_tmpl_id" />
-                </field>
-                <field name="name" position="after">
-                    <field name="product_attribute_ids"
-                           context="{'default_owner_model': 'sale.order.line', 'default_owner_id': id}"
-                           attrs="{'invisible': [('product_attribute_ids', '=', [])]}"
-                    />
-                </field>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="sale_order_line_extended_tree_view">
-            <field name="name">sale.order.line.extended.tree</field>
-            <field name="model">sale.order.line</field>
-            <field name="inherit_id" ref="sale.view_order_line_tree" />
-            <field name="arch" type="xml">
-                <field name="name" position="after">
-                    <field name="product_tmpl_id" />
-                    <field name="product_id" />
-                </field>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="sale_order_line_extended_search_view">
-            <field name="name">sale.order.line.extended.search</field>
-            <field name="model">sale.order.line</field>
-            <field name="inherit_id" ref="sale.view_sales_order_line_filter" />
-            <field name="arch" type="xml">
-                <field name="product_id" position="after">
-                    <field name="product_tmpl_id" />
-                </field>
-                <filter string="Product" position="after">
-                    <filter string="Template" name="search_template"
-                            icon="terp-accessories-archiver"
-                            domain="[]"
-                            context="{'group_by':'product_tmpl_id'}" />
-                </filter>
-            </field>
-        </record>
-    </data>
-</openerp>
