+++ sale_variant_configurator/models/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-# © 2014-2016 Oihane Crucelaegui - AvanzOSC
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
 from . import sale_order
