+++ sale_variant_configurator/views/sale_view.xml
@@ -1,0 +1,66 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <record model="ir.ui.view" id="view_order_form">
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale.view_order_form" />
+        <field name="priority" eval="20" />
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='order_line']/tree" position="attributes">
+                <attribute name="editable"/>
+            </xpath>
+            <xpath expr="//field[@name='order_line']/tree//field[@name='product_id']"
+                   position="before">
+                <field name="product_tmpl_id" />
+            </xpath>
+            <xpath expr="//field[@name='order_line']" position="attributes">
+                <attribute name="options">{'reload_on_button': true}</attribute>
+            </xpath>
+            <xpath expr="//field[@name='order_line']/form//field[@name='product_id']"
+                   position="before">
+                <field name="product_tmpl_id" />
+                <field name="product_attribute_ids"
+                       context="{'default_owner_model': 'sale.order.line', 'default_owner_id': id}"
+                       attrs="{'invisible':[('product_attribute_ids','=',[])]}">
+                    <!-- We are not using isolated view because onchange doesn't work in this case -->
+                    <tree create="0" delete="0" editable="1">
+                        <field name="owner_model" invisible="1"/>
+                        <field name="owner_id" invisible="1"/>
+                        <field name="attribute_id" />
+                        <field name="possible_value_ids" widget="many2many_tags" invisible="1"/>
+                        <field name="product_tmpl_id" invisible="1"/>
+                        <field name="value_id"/>
+                        <field name="price_extra"/>
+                    </tree>
+                </field>
+            </xpath>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="view_order_line_tree">
+        <field name="model">sale.order.line</field>
+        <field name="inherit_id" ref="sale.view_order_line_tree" />
+        <field name="arch" type="xml">
+            <field name="name" position="after">
+                <field name="product_tmpl_id" />
+                <field name="product_id" />
+            </field>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="view_sales_order_line_filter">
+        <field name="model">sale.order.line</field>
+        <field name="inherit_id" ref="sale.view_sales_order_line_filter" />
+        <field name="arch" type="xml">
+            <field name="product_id" position="after">
+                <field name="product_tmpl_id" />
+            </field>
+            <filter context="{'group_by':'product_id'}" position="after">
+                <filter string="Template" name="search_template"
+                        domain="[]"
+                        context="{'group_by':'product_tmpl_id'}" />
+            </filter>
+        </field>
+    </record>
+
+</odoo>
